Ok I found out why theres a bed
APPARENTLY somehow the bed node was an important node despite being called "bed" in a world where a single bed doesn't exist. The bed node contained the popup that said "Press [E] to interact" and all the other code used it. I've removed it and changed the interact menus a bit and now it all works.